Use pickles for caching TOAs
Loading a .tim file and computing Earth positions and velocities can be
time-consuming. PINT now saves the completed TOAs object to disk in a
pickle format in your working directory. If a pickle exists and the
input file is unchanged, PINT will load pre-computed values from the
pickle, which is much faster. Otherwise transparent.